Sound Master App for Windows Phone 8
SoundMaster has a collection of sounds based on categories like animals,
warning etc. It has an additional feature that user can add his sounds
to a separate category.

Note-
There is no option to remove the sound (Use it Carefully).
This app was developed based on the wp8 tutorials on channel9.